Fixed a bug and removed delete option (redundant)
Fixed a bug in the useEffect hook that listened to each key press, it now only receives a set of allowed characters.
This allows use of keys like Tab, Shift, Esc, Return,... without triggering a state change in the game counter.

Removed the backspace button and delete functionality as it was redundant for the game purpose